// ais_mfr_work_time_settings/src/app/work-time-settings/models/Person.model.ts
export interface Person{
    id?: number
    uid: string
    username?: string
    keyloackConnection?: boolean
    name?: string
    name2?:string
    email?: string
    phone?: string
    email2?: string
    phone2?: string
    city?: string
    tnumber?: string
    divisionName?: string
    divisionId:number
    treeDivisionId:number
    post?: string
    status?: string
    status2?: string
    directorDivision?: string
    dateOfEmployment?: Date
    dateOfDismissal?: Date
    avatarId?: number
    keycloakUid:string
  }